+++ src/validateInvocation.test.ts
@@ -21,17 +21,17 @@ describe('#validateInvocation', () => {
     });
 
     await expect(validateInvocation(executionContext)).rejects.toThrow(
-      'Config requires all of {clientId, clientSecret}',
+      'Config requires {apiKey}',
     );
   });
 
   /**
    * Testing a successful authorization can be done with recordings
    */
-  test.skip('successfully validates invocation', async () => {
+  test('successfully validates invocation', async () => {
     recording = setupProjectRecording({
       directory: __dirname,
-      name: 'validate-invocation',
+      name: 'valid-api-key',
     });
 
     // Pass integrationConfig to authenticate with real credentials
@@ -53,11 +53,11 @@ describe('#validateInvocation', () => {
      * For each possible failure case, a test can be made to ensure that
      * error messaging is expected and clear to end-users
      */
-    describe('invalid user credentials', () => {
-      test.skip('should throw if clientId is invalid', async () => {
+    describe('invalid credentials', () => {
+      test('should throw if apiKey is invalid', async () => {
         recording = setupProjectRecording({
           directory: __dirname,
-          name: 'client-id-auth-error',
+          name: 'invalid-api-key',
           // Many authorization failures will return non-200 responses
           // and `recordFailedRequest: true` is needed to capture these responses
           options: {
@@ -67,36 +67,14 @@ describe('#validateInvocation', () => {
 
         const executionContext = createMockExecutionContext({
           instanceConfig: {
-            clientId: 'INVALID',
-            clientSecret: integrationConfig.clientSecret,
+            apiKey: 'INVALID',
           },
         });
 
         // tests validate that invalid configurations throw an error
         // with an appropriate and expected message.
         await expect(validateInvocation(executionContext)).rejects.toThrow(
-          'Provider authentication failed at https://localhost/api/v1/some/endpoint?limit=1: 401 Unauthorized',
-        );
-      });
-
-      test.skip('should throw if clientSecret is invalid', async () => {
-        recording = setupProjectRecording({
-          directory: __dirname,
-          name: 'client-secret-auth-error',
-          options: {
-            recordFailedRequests: true,
-          },
-        });
-
-        const executionContext = createMockExecutionContext({
-          instanceConfig: {
-            clientId: integrationConfig.clientSecret,
-            clientSecret: 'INVALID',
-          },
-        });
-
-        await expect(validateInvocation(executionContext)).rejects.toThrow(
-          'Provider authentication failed at https://localhost/api/v1/some/endpoint?limit=1: 401 Unauthorized',
+          'Provider authentication failed at https://cyberrange.clients.xmcyber.com/api/status/systemHealth: 401 Unauthorized',
         );
       });
     });